requirements.txt
streamlit==1.39.0
tensorflow==2.17.0
mediapipe==0.10.15
opencv-python-headless==4.10.0.84
numpy==1.26.4
pillow==10.4.0
protobuf==4.25.3
